--- src/main/java/com/parkez/reservation/domain/enums/LockStrategy.java
@@ -1,13 +1,0 @@
-package com.parkez.reservation.domain.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum LockStrategy {
-    NONE("noneLockService"),
-    DISTRIBUTED("distributedLockService");
-
-    private final String description;
-}
