--- src/main/java/com/parkez/reservation/service/concurrency/ReservationLockService.java
@@ -1,10 +1,0 @@
-package com.parkez.reservation.service.concurrency;
-
-import com.parkez.common.principal.AuthUser;
-import com.parkez.reservation.dto.request.ReservationRequest;
-import com.parkez.reservation.dto.response.MyReservationResponse;
-
-public interface ReservationLockService {
-
-    MyReservationResponse createReservation(AuthUser authUser, ReservationRequest request);
-}
